Correcting comments and moving 2 functions to utils
